UDNP-59 — Adds eslint-plugin-react
I had to add this plugin because the linter in Atom was complaining that it couldn't find eslint-plugin-react. I tried disabling the plugin in the analytics-api/.eslintrc file by specifying the plugins key as an empty array — that didn't work. I also tried installing and configuring eslint-plugin-disable in the root project, but that didn't work either.

The only thing that worked was installing the plugin in the analytics-api directory. Super lame, but can't spend any more time on this.