UDNP-64 — Fixes Winston file transport config inheritance
Previously, the defaults were taking precedence over the transport-specific configs. Object.assign goes from right to left, with properties in objects on the right overwriting properties in objects on the left. Object.assign also modifies the first object passed to it. In order to avoid modifying the defaults, I added an empty object as the first argument passed to Object.assign.

This commit also renames the log files to accurately reflect the log level (2 == info, 1 == warn).