UDNP-111 — Restructures db.js into a class
This was done to make the code more accessible for tests. I didn't like the old way I set the file up with a bunch of private functions that had to be manually added to the db object.

Now it just exports an instance of the AnalyticsDB class — much cleaner.