Refactor validator
Instead of having a validation function for each type, I made an object of types that each have a message function and a validator regex. This eliminates duplicated logic in each of the validation functions.

There is now simply a _validateValue function that references the _types object to get the regex and message.

Also added two new types: service and granularity.

Also updated the units to reflect the changes/additions.