Redo user menu structure for styling
Previous implementation didn't work well if the user menu dropdown was taller than browser window. Had to add extra nested ul > li wrappers for the 'normal' list items and use sticky footer method to get the last Log Out link to be at the bottom of the dropdown. With this method the Log Out link doesn't overflow on top of the other items if the dropdown is taller than browser window.