Remove group resetting
Is this ok? Otherwise the active group is reset when user click the group link in breadcrumb header. Then if they use browser back button to get back to property page the active group name is missing from breadcrumbs. The breadcrumb link can still be formed correctly using params, but not the name, as params only stores ids.