Change reset password flow
Added inputs to set new password and to confirm it. Added validation to check that password and confirm match, if new password has been entered. Showing the error only if neither input has focus. Modified save function to check if new password has been set before so that we don’t try to send empty password to API, which would cause an error.