Fix setting startDate and endDate
If user has selected some date range and navigates away from the page and then comes back, the date range selector wouldn’t be updated as it still had in the redux state the old date range. Somehow the function to reset the filters has a slight delay and it is completed only after this component has already been mounted. Therefore we have this extra check in componentWillReceiveProps event to check if the next date range props are different from current, and we update them in the component state. TODO: Not working on property summary page