UDNP-1795 — Fix 404 errors when requesting role permissions
We are now getting permissions for each role by requesting the /roles/{id}/services endpoints which return permissions for each service for a given role. Previosuly we were manually requesting permissions for each role for the AAA, North and UI services. The problem was that the new CIS roles had no permissions for the North service, which caused 404 errors when trying to request roles/{id}/services/North.