UDNP-1827 — Fix visitor avg on property summary
The property summary used to show the total number of users returned by the visitors/country endpoint in the analytics API. That was incorrect because the number is supposed to be the average number of unique visitors per hour.

This change now fetches visitor data by time and calculates the average in the reducer.