Make MultiMonthView rendering conditional to DD being open
Apparently new React-Bootstrap has changed the way they render hidden dropdowns and MultiMonthView needs a complete re-rendering in order for the defaultRange prop to be re-initiated. Added condition to render it only when the dropdown is open, so that we can force the re-rendering and make the range updates work again.