UDNP-1436 — Use RegExp instead of substring
Now uses a RegExp to get the list of file extensions out of a RegExp string, instead of using substring to do the same thing.

Also modified the extensionsFromMatchCase function to return an empty array if nothing was found in the input string. An empty array is the proper “empty” value for the tokenizable input field.